Wrap sign-in page with Suspense
